refactor: enhance grimoire state management by updating function parameters and consolidating imports across modules
