+++ script.js
@@ -1,14 +1,14 @@
-"use strict";
 import './pwa.js';
-import { addGrimoireHistoryListListeners, renderGrimoireHistory, snapshotCurrentGrimoire } from "./ui/history/grimoire.js";
+import { addGrimoireHistoryListListeners, renderGrimoireHistory } from "./ui/history/grimoire.js";
 import { loadHistories } from './ui/history/index.js';
 import { addScriptHistoryListListeners, addScriptToHistory, renderScriptHistory } from "./ui/history/script.js";
 import { positionRadialStack as positionRadialStackLayout, repositionPlayers as repositionPlayersLayout } from './ui/layout.js';
 import { initSidebarResize, initSidebarToggle } from './ui/sidebar.js';
-import { createCurvedLabelSvg, createDeathRibbonSvg } from './ui/svg.js';
-import { positionInfoIcons, positionTooltip, showTouchAbilityPopup } from './ui/tooltip.js';
+import { createCurvedLabelSvg } from './ui/svg.js';
 import { initInAppTour } from './ui/tour.js';
 import { isExcludedScriptName, normalizeKey, resolveAssetPath } from './utils.js';
+import { saveAppState, renderSetupInfo, loadAppState, startGame, updateGrimoire } from './ui/grimoire.js';
+import { CLICK_EXPAND_SUPPRESS_MS, TOUCH_EXPAND_SUPPRESS_MS, isTouchDevice, INCLUDE_TRAVELLERS_KEY } from './constants.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   const startGameBtn = document.getElementById('start-game');
@@ -19,7 +19,6 @@ document.addEventListener('DOMContentLoaded', () => {
   const scriptFileInput = document.getElementById('script-file');
   const playerCountInput = document.getElementById('player-count');
   const playerCircle = document.getElementById('player-circle');
-  const setupInfoEl = document.getElementById('setup-info');
   const characterSheet = document.getElementById('character-sheet');
   const loadStatus = document.getElementById('load-status');
 
@@ -36,7 +35,6 @@ document.addEventListener('DOMContentLoaded', () => {
   const sidebarResizer = document.getElementById('sidebar-resizer');
 
   // Ability tooltip elements
-  const abilityTooltip = document.getElementById('ability-tooltip');
   const sidebarEl = document.getElementById('sidebar');
   const reminderTokenModal = document.getElementById('reminder-token-modal');
   const closeReminderTokenModalBtn = document.getElementById('close-reminder-token-modal');
@@ -53,7 +51,6 @@ document.addEventListener('DOMContentLoaded', () => {
   const centerEl = document.getElementById('center');
   const includeTravellersCheckbox = document.getElementById('include-travellers');
   // Travellers toggle state key and default
-  const INCLUDE_TRAVELLERS_KEY = 'botcIncludeTravellersV1';
 
   const grimoireState = {
     includeTravellers: false,
@@ -115,17 +112,13 @@ document.addEventListener('DOMContentLoaded', () => {
     includeTravellersCheckbox.addEventListener('change', () => {
       grimoireState.includeTravellers = !!includeTravellersCheckbox.checked;
       applyTravellerToggleAndRefresh();
-      saveAppState();
+      saveAppState({ grimoireState });
     });
   }
 
-  const isTouchDevice = (window.matchMedia && window.matchMedia('(hover: none) and (pointer: coarse)').matches) || ('ontouchstart' in window) || (navigator.maxTouchPoints > 0);
-  const TOUCH_EXPAND_SUPPRESS_MS = 350;
-  const CLICK_EXPAND_SUPPRESS_MS = 250;
-  const prefersOverlaySidebar = window.matchMedia('(max-width: 900px)');
-
   // Build player circle UI from current players WITHOUT snapshotting or resetting players
   function rebuildPlayerCircleUiPreserveState() {
+    const playerCircle = document.getElementById('player-circle');
     if (!playerCircle) return;
     playerCircle.innerHTML = '';
     // Keep sidebar input in sync with current number of players
@@ -159,8 +152,8 @@ document.addEventListener('DOMContentLoaded', () => {
         const newName = prompt("Enter player name:", player.name);
         if (newName) {
           grimoireState.players[i].name = newName;
-          updateGrimoire();
-          saveAppState();
+          updateGrimoire({ grimoireState });
+          saveAppState({ grimoireState });
         }
       };
       listItem.querySelector('.reminder-placeholder').onclick = (e) => {
@@ -280,14 +273,14 @@ document.addEventListener('DOMContentLoaded', () => {
       }
     });
     // Apply layout and state immediately for deterministic testing and UX
-    repositionPlayers();
-    updateGrimoire();
-    saveAppState();
-    renderSetupInfo();
+    repositionPlayers(grimoireState.players);
+    updateGrimoire({ grimoireState });
+    saveAppState({ grimoireState });
+    renderSetupInfo({ grimoireState });
     // Also after paint to ensure positions stabilize
     requestAnimationFrame(() => {
-      repositionPlayers();
-      updateGrimoire();
+      repositionPlayers(grimoireState.players);
+      updateGrimoire({ grimoireState });
     });
   }
 
@@ -406,8 +399,8 @@ document.addEventListener('DOMContentLoaded', () => {
           rem.value = next;
           if (rem.label !== undefined) rem.label = next;
         }
-        updateGrimoire();
-        saveAppState();
+        updateGrimoire({ grimoireState });
+        saveAppState({ grimoireState });
       }
     });
 
@@ -417,8 +410,8 @@ document.addEventListener('DOMContentLoaded', () => {
       if (playerIndex < 0 || reminderIndex < 0) return;
       if (!grimoireState.players[playerIndex] || !grimoireState.players[playerIndex].reminders) return;
       grimoireState.players[playerIndex].reminders.splice(reminderIndex, 1);
-      updateGrimoire();
-      saveAppState();
+      updateGrimoire({ grimoireState });
+      saveAppState({ grimoireState });
     });
 
     menu.appendChild(editBtn);
@@ -467,43 +460,12 @@ document.addEventListener('DOMContentLoaded', () => {
 
   // Event delegation for history lists
   if (scriptHistoryList) {
-    addScriptHistoryListListeners({ scriptHistoryList, processScriptData, displayScript, saveAppState, renderSetupInfo });
+    addScriptHistoryListListeners({ scriptHistoryList, processScriptData, displayScript });
   }
 
   if (grimoireHistoryList) {
-    addGrimoireHistoryListListeners({ grimoireHistoryList, grimoireState, processScriptData, setupGrimoire, updateGrimoire, repositionPlayers, saveAppState, renderSetupInfo });
+    addGrimoireHistoryListListeners({ grimoireHistoryList, grimoireState, openCharacterModal, showPlayerContextMenu, openReminderTokenModal, openTextReminderModal, processScriptData, repositionPlayers });
   }
-
-  function saveAppState() {
-    try {
-      const state = { scriptData: grimoireState.scriptData, players: grimoireState.players, scriptName: grimoireState.scriptMetaName };
-      localStorage.setItem('botcAppStateV1', JSON.stringify(state));
-      try { localStorage.setItem(INCLUDE_TRAVELLERS_KEY, grimoireState.includeTravellers ? '1' : '0'); } catch (_) { }
-    } catch (_) { }
-  }
-
-  async function loadAppState() {
-    try {
-      grimoireState.isRestoringState = true;
-      const raw = localStorage.getItem('botcAppStateV1');
-      if (!raw) return;
-      const saved = JSON.parse(raw);
-      if (saved && Array.isArray(saved.scriptData) && saved.scriptData.length) {
-        await processScriptData(saved.scriptData, false);
-        if (saved.scriptMetaName) { grimoireState.scriptMetaName = String(saved.scriptMetaName); }
-        if (saved.includeTravellers) { grimoireState.includeTravellers = saved.includeTravellers; }
-      }
-      if (saved && Array.isArray(saved.players) && saved.players.length) {
-        setupGrimoire(saved.players.length);
-        grimoireState.players = saved.players;
-        updateGrimoire();
-        repositionPlayers();
-        renderSetupInfo();
-      }
-    } catch (_) { } finally { grimoireState.isRestoringState = false; }
-  }
-
-  // Path and key helpers imported from utils.js
 
   async function loadAllCharacters() {
     try {
@@ -549,7 +511,7 @@ document.addEventListener('DOMContentLoaded', () => {
       grimoireState.scriptData = [{ id: '_meta', name: 'All Characters', author: 'System' }, ...characterIds];
       // Apply traveller toggle to compute allRoles and render
       applyTravellerToggleAndRefresh();
-      saveAppState();
+      saveAppState({ grimoireState });
 
       loadStatus.textContent = `Loaded ${Object.keys(grimoireState.allRoles).length} characters successfully`;
       loadStatus.className = 'status';
@@ -571,7 +533,7 @@ document.addEventListener('DOMContentLoaded', () => {
         if (match) {
           const base = match[1].replace(/\s*&\s*/g, ' & ');
           grimoireState.scriptMetaName = base;
-          renderSetupInfo();
+          renderSetupInfo({ grimoireState });
         }
       } catch (_) { }
       const res = await fetch(path, { cache: 'no-store' });
@@ -586,10 +548,10 @@ document.addEventListener('DOMContentLoaded', () => {
       loadStatus.className = 'error';
     }
   }
-  loadTbBtn && loadTbBtn.addEventListener('click', () => { grimoireState.scriptMetaName = 'Trouble Brewing'; renderSetupInfo(); loadScriptFromFile('./Trouble Brewing.json'); });
-  loadBmrBtn && loadBmrBtn.addEventListener('click', () => { grimoireState.scriptMetaName = 'Bad Moon Rising'; renderSetupInfo(); loadScriptFromFile('./Bad Moon Rising.json'); });
-  loadSavBtn && loadSavBtn.addEventListener('click', () => { grimoireState.scriptMetaName = 'Sects & Violets'; renderSetupInfo(); loadScriptFromFile('./Sects and Violets.json'); });
-  loadAllCharsBtn && loadAllCharsBtn.addEventListener('click', () => { grimoireState.scriptMetaName = 'All Characters'; renderSetupInfo(); loadAllCharacters(); });
+  loadTbBtn && loadTbBtn.addEventListener('click', () => { grimoireState.scriptMetaName = 'Trouble Brewing'; renderSetupInfo({ grimoireState }); loadScriptFromFile('./Trouble Brewing.json'); });
+  loadBmrBtn && loadBmrBtn.addEventListener('click', () => { grimoireState.scriptMetaName = 'Bad Moon Rising'; renderSetupInfo({ grimoireState }); loadScriptFromFile('./Bad Moon Rising.json'); });
+  loadSavBtn && loadSavBtn.addEventListener('click', () => { grimoireState.scriptMetaName = 'Sects & Violets'; renderSetupInfo({ grimoireState }); loadScriptFromFile('./Sects and Violets.json'); });
+  loadAllCharsBtn && loadAllCharsBtn.addEventListener('click', () => { grimoireState.scriptMetaName = 'All Characters'; renderSetupInfo({ grimoireState }); loadAllCharacters(); });
 
   scriptFileInput.addEventListener('change', async (event) => {
     const file = event.target.files[0];
@@ -630,34 +592,11 @@ document.addEventListener('DOMContentLoaded', () => {
       if (!res.ok) throw new Error(`HTTP ${res.status}`);
       const data = await res.json();
       grimoireState.playerSetupTable = Array.isArray(data.player_setup) ? data.player_setup : [];
-      renderSetupInfo();
+      renderSetupInfo({ grimoireState });
     } catch (e) {
       console.error('Failed to load player-setup.json', e);
     }
   })();
-
-  function lookupCountsForPlayers(count) {
-    if (!Array.isArray(grimoireState.playerSetupTable)) return null;
-    const row = grimoireState.playerSetupTable.find(r => Number(r.players) === Number(count));
-    return row || null;
-  }
-
-  function renderSetupInfo() {
-    if (!setupInfoEl) return;
-    const count = grimoireState.players.length;
-    const row = lookupCountsForPlayers(count);
-    // Prefer parsed meta name; otherwise keep any existing hint
-    let scriptName = grimoireState.scriptMetaName || '';
-    if (!scriptName && Array.isArray(grimoireState.scriptData)) {
-      const meta = grimoireState.scriptData.find(x => x && typeof x === 'object' && x.id === '_meta');
-      if (meta && meta.name) scriptName = String(meta.name);
-    }
-    if (!row && !scriptName) { setupInfoEl.textContent = 'Select a script and add players from the sidebar.'; return; }
-    const parts = [];
-    if (scriptName) parts.push(scriptName);
-    if (row) parts.push(`${row.townsfolk}/${row.outsiders}/${row.minions}/${row.demons}`);
-    setupInfoEl.textContent = parts.join(' ');
-  }
 
   async function processScriptData(data, addToHistory = false) {
     console.log('Processing script data:', data);
@@ -682,8 +621,8 @@ document.addEventListener('DOMContentLoaded', () => {
     console.log('Total roles processed:', Object.keys(grimoireState.allRoles).length);
     // After processing into baseRoles/extraTravellerRoles, apply toggle
     applyTravellerToggleAndRefresh();
-    saveAppState();
-    renderSetupInfo();
+    saveAppState({ grimoireState });
+    renderSetupInfo({ grimoireState });
     if (addToHistory) {
       const histName = grimoireState.scriptMetaName || (Array.isArray(data) && (data.find(x => x && typeof x === 'object' && x.id === '_meta')?.name || 'Custom Script')) || 'Custom Script';
       if (!isExcludedScriptName(histName)) {
@@ -813,599 +752,9 @@ document.addEventListener('DOMContentLoaded', () => {
     }
   }
 
-  startGameBtn.addEventListener('click', () => {
-    const playerCount = parseInt(playerCountInput.value, 10);
-    if (playerCount >= 5 && playerCount <= 20) {
-      setupGrimoire(playerCount);
-    } else {
-      alert('Player count must be an integer from 5 to 20.');
-    }
-  });
-
-  function setupGrimoire(count) {
-    try {
-      if (!grimoireState.isRestoringState && Array.isArray(grimoireState.players) && grimoireState.players.length > 0) {
-        snapshotCurrentGrimoire({ players: grimoireState.players, scriptMetaName: grimoireState.scriptMetaName, scriptData: grimoireState.scriptData, grimoireHistoryList });
-      }
-    } catch (_) { }
-    console.log('Setting up grimoire with', count, 'players');
-    playerCircle.innerHTML = '';
-    grimoireState.players = Array.from({ length: count }, (_, i) => ({
-      name: `Player ${i + 1}`,
-      character: null,
-      reminders: [],
-      dead: false
-    }));
-
-    grimoireState.players.forEach((player, i) => {
-      const listItem = document.createElement('li');
-      listItem.innerHTML = `
-              <div class="reminders"></div>
-              <div class="player-token" title="Assign character"></div>
-               <div class="character-name" aria-live="polite"></div>
-              <div class="player-name" title="Edit name">${player.name}</div>
-              <div class="reminder-placeholder" title="Add text reminder">+</div>
-          `;
-      playerCircle.appendChild(listItem);
-
-      // Only the main token area opens the character modal; ribbon handles dead toggle
-      listItem.querySelector('.player-token').onclick = (e) => {
-        const target = e.target;
-        if (target && (target.closest('.death-ribbon') || target.classList.contains('death-ribbon'))) {
-          return; // handled by ribbon click
-        }
-        if (target && target.classList.contains('ability-info-icon')) {
-          return; // handled by info icon
-        }
-        openCharacterModal(i);
-      };
-      // Player context menu: right-click
-      listItem.addEventListener('contextmenu', (e) => {
-        e.preventDefault();
-        showPlayerContextMenu(e.clientX, e.clientY, i);
-      });
-      // Long-press on token to open context menu on touch devices
-      const tokenForMenu = listItem.querySelector('.player-token');
-      tokenForMenu.addEventListener('pointerdown', (e) => {
-        if (!isTouchDevice) return;
-        try { e.preventDefault(); } catch (_) { }
-        clearTimeout(grimoireState.longPressTimer);
-        const x = (e && (e.clientX || (e.touches && e.touches[0] && e.touches[0].clientX))) || 0;
-        const y = (e && (e.clientY || (e.touches && e.touches[0] && e.touches[0].clientY))) || 0;
-        grimoireState.longPressTimer = setTimeout(() => {
-          showPlayerContextMenu(x, y, i);
-        }, 600);
-      });
-      ['pointerup', 'pointercancel', 'pointerleave'].forEach(evt => {
-        tokenForMenu.addEventListener(evt, () => { clearTimeout(grimoireState.longPressTimer); });
-      });
-      listItem.querySelector('.player-name').onclick = (e) => {
-        e.stopPropagation();
-        const newName = prompt("Enter player name:", player.name);
-        if (newName) {
-          grimoireState.players[i].name = newName;
-          updateGrimoire();
-          saveAppState();
-        }
-      };
-      listItem.querySelector('.reminder-placeholder').onclick = (e) => {
-        e.stopPropagation();
-        const thisLi = listItem;
-        if (thisLi.dataset.expanded !== '1') {
-          const allLis = document.querySelectorAll('#player-circle li');
-          let someoneExpanded = false;
-          allLis.forEach(el => {
-            if (el !== thisLi && el.dataset.expanded === '1') {
-              someoneExpanded = true;
-              el.dataset.expanded = '0';
-              const idx = Array.from(allLis).indexOf(el);
-              positionRadialStack(el, (grimoireState.players[idx]?.reminders || []).length);
-            }
-          });
-          if (someoneExpanded) {
-            thisLi.dataset.expanded = '1';
-            thisLi.dataset.actionSuppressUntil = String(Date.now() + CLICK_EXPAND_SUPPRESS_MS);
-            positionRadialStack(thisLi, grimoireState.players[i].reminders.length);
-            return;
-          }
-        }
-        if (isTouchDevice) {
-          openReminderTokenModal(i);
-        } else if (e.altKey) {
-          openTextReminderModal(i);
-        } else {
-          openReminderTokenModal(i);
-        }
-      };
-
-      // Hover expand/collapse for reminder stack positioning
-      listItem.dataset.expanded = '0';
-      const expand = () => {
-        const wasExpanded = listItem.dataset.expanded === '1';
-        const allLis = document.querySelectorAll('#player-circle li');
-        allLis.forEach(el => {
-          if (el !== listItem && el.dataset.expanded === '1') {
-            el.dataset.expanded = '0';
-            const idx = Array.from(allLis).indexOf(el);
-            positionRadialStack(el, (grimoireState.players[idx]?.reminders || []).length);
-          }
-        });
-        listItem.dataset.expanded = '1';
-        // Only set suppression on touch, and only when changing from collapsed -> expanded
-        if (isTouchDevice && !wasExpanded) {
-          listItem.dataset.actionSuppressUntil = String(Date.now() + CLICK_EXPAND_SUPPRESS_MS);
-        }
-        positionRadialStack(listItem, grimoireState.players[i].reminders.length);
-      };
-      const collapse = () => { listItem.dataset.expanded = '0'; positionRadialStack(listItem, grimoireState.players[i].reminders.length); };
-      if (!isTouchDevice) {
-        listItem.addEventListener('mouseenter', expand);
-        listItem.addEventListener('mouseleave', collapse);
-        // Pointer events for broader device support
-        listItem.addEventListener('pointerenter', expand);
-        listItem.addEventListener('pointerleave', collapse);
-      }
-      // Touch: expand on any tap; only suppress synthetic click if tap started on reminders
-      listItem.addEventListener('touchstart', (e) => {
-        const target = e.target;
-        const tappedReminders = !!(target && target.closest('.reminders'));
-        if (tappedReminders) {
-          try { e.preventDefault(); } catch (_) { }
-          listItem.dataset.touchSuppressUntil = String(Date.now() + TOUCH_EXPAND_SUPPRESS_MS);
-        }
-        expand();
-        positionRadialStack(listItem, grimoireState.players[i].reminders.length);
-      }, { passive: false });
-
-      // (desktop) no extra mousedown handler; rely on hover/pointerenter and explicit clicks on reminders
-
-      // Install one-time outside click/tap collapse for touch devices
-      if (isTouchDevice && !grimoireState.outsideCollapseHandlerInstalled) {
-        grimoireState.outsideCollapseHandlerInstalled = true;
-        const maybeCollapseOnOutside = (ev) => {
-          const target = ev.target;
-          // Ignore clicks/taps inside the player circle to allow in-circle interactions (like + gating)
-          const playerCircleEl = document.getElementById('player-circle');
-          if (playerCircleEl && playerCircleEl.contains(target)) return;
-          // Do nothing if target is inside any expanded list item
-          const allLis = document.querySelectorAll('#player-circle li');
-          let clickedInsideExpanded = false;
-          allLis.forEach(el => {
-            if (el.dataset.expanded === '1' && el.contains(target)) {
-              clickedInsideExpanded = true;
-            }
-          });
-          if (clickedInsideExpanded) return;
-          // Collapse all expanded items
-          allLis.forEach(el => {
-            if (el.dataset.expanded === '1') {
-              el.dataset.expanded = '0';
-              positionRadialStack(el, (grimoireState.players[Array.from(allLis).indexOf(el)]?.reminders || []).length);
-            }
-          });
-        };
-        document.addEventListener('click', maybeCollapseOnOutside, true);
-        document.addEventListener('touchstart', maybeCollapseOnOutside, { passive: true, capture: true });
-      }
-
-      // No capture intercepts; rely on pointer-events gating and the touchstart handler above
-    });
-
-    // Use requestAnimationFrame to ensure DOM is fully rendered
-    requestAnimationFrame(() => {
-      repositionPlayers();
-      updateGrimoire();
-      saveAppState();
-      renderSetupInfo();
-    });
-  }
+  startGameBtn.addEventListener('click', () => startGame({ grimoireState, grimoireHistoryList, playerCountInput, openCharacterModal, showPlayerContextMenu, openReminderTokenModal, openTextReminderModal }));
 
   function repositionPlayers() { repositionPlayersLayout(grimoireState.players); }
-
-  function updateGrimoire() {
-    const listItems = playerCircle.querySelectorAll('li');
-    listItems.forEach((li, i) => {
-      const player = grimoireState.players[i];
-      const playerNameEl = li.querySelector('.player-name');
-      playerNameEl.textContent = player.name;
-
-      // Check if player is in NW or NE quadrant
-      const angle = parseFloat(li.dataset.angle || '0');
-
-      // Calculate the actual x,y position to determine quadrant
-      const x = Math.cos(angle);
-      const y = Math.sin(angle);
-
-      // Names go on top for NW (x<0, y<0) and NE (x>0, y<0) quadrants
-      const isNorthQuadrant = y < 0;
-
-      if (isNorthQuadrant) {
-        playerNameEl.classList.add('top-half');
-        li.classList.add('is-north');
-        li.classList.remove('is-south');
-      } else {
-        playerNameEl.classList.remove('top-half');
-        li.classList.add('is-south');
-        li.classList.remove('is-north');
-      }
-
-      const tokenDiv = li.querySelector('.player-token');
-      const charNameDiv = li.querySelector('.character-name');
-      // Remove any previous arc label overlay
-      const existingArc = tokenDiv.querySelector('.icon-reminder-svg');
-      if (existingArc) existingArc.remove();
-      // Remove any previous death UI
-      const oldCircle = tokenDiv.querySelector('.death-overlay');
-      if (oldCircle) oldCircle.remove();
-      const oldRibbon = tokenDiv.querySelector('.death-ribbon');
-      if (oldRibbon) oldRibbon.remove();
-
-      if (player.character) {
-        const role = getRoleById(player.character);
-        if (role) {
-          tokenDiv.style.backgroundImage = `url('${resolveAssetPath(role.image)}'), url('${resolveAssetPath('assets/img/token-BqDQdWeO.webp')}')`;
-          tokenDiv.style.backgroundSize = '68% 68%, cover';
-          tokenDiv.style.backgroundColor = 'transparent';
-          tokenDiv.classList.add('has-character');
-          if (charNameDiv) charNameDiv.textContent = role.name;
-          // Add curved label on the token
-          const svg = createCurvedLabelSvg(`player-arc-${i}`, role.name);
-          tokenDiv.appendChild(svg);
-
-          // Add tooltip functionality for non-touch devices
-          if (!('ontouchstart' in window)) {
-            tokenDiv.addEventListener('mouseenter', (e) => {
-              if (role.ability) {
-                abilityTooltip.textContent = role.ability;
-                abilityTooltip.classList.add('show');
-                positionTooltip(e.target, abilityTooltip);
-              }
-            });
-
-            tokenDiv.addEventListener('mouseleave', () => {
-              abilityTooltip.classList.remove('show');
-            });
-          } else if (role.ability) {
-            // Add info icon for touch mode - will be positioned after circle layout
-            const infoIcon = document.createElement('div');
-            infoIcon.className = 'ability-info-icon';
-            infoIcon.innerHTML = '<i class="fas fa-info-circle"></i>';
-            infoIcon.dataset.playerIndex = i;
-            // Handle both click and touch events
-            const handleInfoClick = (e) => {
-              e.stopPropagation();
-              e.preventDefault();
-              showTouchAbilityPopup(infoIcon, role.ability);
-            };
-            infoIcon.onclick = handleInfoClick;
-            infoIcon.addEventListener('touchstart', (e) => {
-              e.stopPropagation();
-              e.preventDefault();
-              handleInfoClick(e); // Call the click handler on touch
-            });
-            li.appendChild(infoIcon); // Append to li, not tokenDiv
-          }
-        } else {
-          tokenDiv.style.backgroundImage = `url('${resolveAssetPath('assets/img/token-BqDQdWeO.webp')}')`;
-          tokenDiv.style.backgroundSize = 'cover';
-          tokenDiv.style.backgroundColor = 'rgba(0,0,0,0.2)';
-          tokenDiv.classList.remove('has-character');
-          if (charNameDiv) charNameDiv.textContent = '';
-          // Ensure no leftover arc label remains
-          const arc = tokenDiv.querySelector('.icon-reminder-svg');
-          if (arc) arc.remove();
-        }
-      } else {
-        tokenDiv.style.backgroundImage = `url('${resolveAssetPath('assets/img/token-BqDQdWeO.webp')}')`;
-        tokenDiv.style.backgroundSize = 'cover';
-        tokenDiv.style.backgroundColor = 'rgba(0,0,0,0.2)';
-        tokenDiv.classList.remove('has-character');
-        if (charNameDiv) charNameDiv.textContent = '';
-        // Ensure no leftover arc label remains
-        const arc = tokenDiv.querySelector('.icon-reminder-svg');
-        if (arc) arc.remove();
-      }
-
-      // Add death overlay circle and ribbon indicator
-      const overlay = document.createElement('div');
-      overlay.className = 'death-overlay';
-      overlay.title = player.dead ? 'Click to mark alive' : 'Click to mark dead';
-      // overlay is visual only; click is on ribbon
-      tokenDiv.appendChild(overlay);
-
-      const ribbon = createDeathRibbonSvg();
-      ribbon.classList.add('death-ribbon');
-      const handleRibbonToggle = (e) => {
-        e.stopPropagation();
-        grimoireState.players[i].dead = !grimoireState.players[i].dead;
-        updateGrimoire();
-        saveAppState();
-      };
-      // Attach to painted shapes only to avoid transparent hit areas
-      try {
-        ribbon.querySelectorAll('rect, path').forEach((shape) => {
-          shape.addEventListener('click', handleRibbonToggle);
-        });
-      } catch (_) {
-        // Fallback: still attach on svg
-        ribbon.addEventListener('click', handleRibbonToggle);
-      }
-      tokenDiv.appendChild(ribbon);
-
-      if (player.dead) {
-        tokenDiv.classList.add('is-dead');
-      } else {
-        tokenDiv.classList.remove('is-dead');
-      }
-
-      const remindersDiv = li.querySelector('.reminders');
-      remindersDiv.innerHTML = '';
-
-      // Create reminder elements; positions are handled by positionRadialStack()
-      player.reminders.forEach((reminder, idx) => {
-        if (reminder.type === 'icon') {
-          const iconEl = document.createElement('div');
-          iconEl.className = 'icon-reminder';
-          iconEl.style.transform = `translate(-50%, -50%) rotate(${reminder.rotation || 0}deg)`;
-          iconEl.style.backgroundImage = `url('${resolveAssetPath(reminder.image)}'), url('${resolveAssetPath('assets/img/token-BqDQdWeO.webp')}')`;
-          iconEl.title = (reminder.label || '');
-          iconEl.addEventListener('click', (e) => {
-            const parentLi = iconEl.closest('li');
-            const isCollapsed = !!(parentLi && parentLi.dataset.expanded !== '1');
-            if (isCollapsed) {
-              e.stopPropagation();
-              try { e.preventDefault(); } catch (_) { }
-              parentLi.dataset.expanded = '1';
-              parentLi.dataset.actionSuppressUntil = String(Date.now() + CLICK_EXPAND_SUPPRESS_MS);
-              positionRadialStack(parentLi, grimoireState.players[i].reminders.length);
-            }
-          }, true);
-
-          if (reminder.label) {
-            // Check if this is a custom reminder by ID
-            const isCustom = reminder.id === 'custom-note';
-
-            if (isCustom) {
-              // For custom reminders, show straight text with dark background
-              const textSpan = document.createElement('span');
-              textSpan.className = 'icon-reminder-content';
-              textSpan.textContent = reminder.label;
-
-              // Adjust font size based on text length
-              const textLength = reminder.label.length;
-              if (textLength > 40) {
-                textSpan.style.fontSize = 'clamp(7px, calc(var(--token-size) * 0.06), 10px)';
-              } else if (textLength > 20) {
-                textSpan.style.fontSize = 'clamp(8px, calc(var(--token-size) * 0.07), 12px)';
-              }
-
-              iconEl.appendChild(textSpan);
-            } else {
-              // For other reminders, show curved text at bottom
-              const svg = createCurvedLabelSvg(`arc-${i}-${idx}`, reminder.label);
-              iconEl.appendChild(svg);
-            }
-          }
-
-          // Desktop hover actions on icon reminders
-          if (!isTouchDevice) {
-            const editBtn = document.createElement('div');
-            editBtn.className = 'reminder-action edit';
-            editBtn.title = 'Edit';
-            editBtn.innerHTML = '<i class="fa-solid fa-pen"></i>';
-            editBtn.addEventListener('click', (e) => {
-              e.stopPropagation();
-              try { e.preventDefault(); } catch (_) { }
-              const parentLi = editBtn.closest('li');
-              if (parentLi) {
-                const suppressUntil = parseInt(parentLi.dataset.actionSuppressUntil || '0', 10);
-                if (parentLi.dataset.expanded !== '1' || Date.now() < suppressUntil) {
-                  if (parentLi.dataset.expanded !== '1') {
-                    parentLi.dataset.expanded = '1';
-                    parentLi.dataset.actionSuppressUntil = String(Date.now() + CLICK_EXPAND_SUPPRESS_MS);
-                    positionRadialStack(parentLi, grimoireState.players[i].reminders.length);
-                  }
-                  return;
-                }
-              }
-              const current = grimoireState.players[i].reminders[idx]?.label || grimoireState.players[i].reminders[idx]?.value || '';
-              const next = prompt('Edit reminder', current);
-              if (next !== null) {
-                grimoireState.players[i].reminders[idx].label = next;
-                updateGrimoire();
-                saveAppState();
-              }
-            });
-            iconEl.appendChild(editBtn);
-
-            const delBtn = document.createElement('div');
-            delBtn.className = 'reminder-action delete';
-            delBtn.title = 'Delete';
-            delBtn.innerHTML = '<i class="fa-solid fa-trash"></i>';
-            delBtn.addEventListener('click', (e) => {
-              e.stopPropagation();
-              try { e.preventDefault(); } catch (_) { }
-              const parentLi = delBtn.closest('li');
-              if (parentLi) {
-                const suppressUntil = parseInt(parentLi.dataset.actionSuppressUntil || '0', 10);
-                if (parentLi.dataset.expanded !== '1' || Date.now() < suppressUntil) {
-                  if (parentLi.dataset.expanded !== '1') {
-                    parentLi.dataset.expanded = '1';
-                    parentLi.dataset.actionSuppressUntil = String(Date.now() + CLICK_EXPAND_SUPPRESS_MS);
-                    positionRadialStack(parentLi, grimoireState.players[i].reminders.length);
-                  }
-                  return;
-                }
-              }
-              grimoireState.players[i].reminders.splice(idx, 1);
-              updateGrimoire();
-              saveAppState();
-            });
-            iconEl.appendChild(delBtn);
-          }
-
-          // Touch long-press for reminder context menu (iOS Safari, Android)
-          if (isTouchDevice) {
-            const onPressStart = (e) => {
-              try { e.preventDefault(); } catch (_) { }
-              clearTimeout(grimoireState.longPressTimer);
-              try { iconEl.classList.add('press-feedback'); } catch (_) { }
-              const x = (e && (e.clientX || (e.touches && e.touches[0] && e.touches[0].clientX))) || 0;
-              const y = (e && (e.clientY || (e.touches && e.touches[0] && e.touches[0].clientY))) || 0;
-              grimoireState.longPressTimer = setTimeout(() => {
-                try { iconEl.classList.remove('press-feedback'); } catch (_) { }
-                showReminderContextMenu(x, y, i, idx);
-              }, 600);
-            };
-            const onPressEnd = () => { clearTimeout(grimoireState.longPressTimer); try { iconEl.classList.remove('press-feedback'); } catch (_) { } };
-            iconEl.addEventListener('pointerdown', onPressStart);
-            iconEl.addEventListener('pointerup', onPressEnd);
-            iconEl.addEventListener('pointercancel', onPressEnd);
-            iconEl.addEventListener('pointerleave', onPressEnd);
-            iconEl.addEventListener('touchstart', onPressStart, { passive: false });
-            iconEl.addEventListener('touchend', onPressEnd);
-            iconEl.addEventListener('touchcancel', onPressEnd);
-            iconEl.addEventListener('contextmenu', (e) => { try { e.preventDefault(); } catch (_) { } });
-          }
-
-          remindersDiv.appendChild(iconEl);
-        } else {
-          const reminderEl = document.createElement('div');
-          reminderEl.className = 'text-reminder';
-
-          // Check if this is actually a text reminder with a label (legacy data)
-          // If so, use the label as the display text
-          const displayText = reminder.label || reminder.value || '';
-
-          // Create a span for the text with dark background
-          const textSpan = document.createElement('span');
-          textSpan.className = 'text-reminder-content';
-          textSpan.textContent = displayText;
-
-          // Adjust font size based on text length
-          const textLength = displayText.length;
-          if (textLength > 40) {
-            textSpan.style.fontSize = 'clamp(7px, calc(var(--token-size) * 0.06), 10px)';
-          } else if (textLength > 20) {
-            textSpan.style.fontSize = 'clamp(8px, calc(var(--token-size) * 0.07), 12px)';
-          }
-
-          reminderEl.appendChild(textSpan);
-
-          reminderEl.style.transform = 'translate(-50%, -50%)';
-          reminderEl.onclick = (e) => {
-            e.stopPropagation();
-            const parentLi = reminderEl.closest('li');
-            if (parentLi) {
-              const suppressUntil = parseInt(parentLi.dataset.actionSuppressUntil || '0', 10);
-              if (parentLi.dataset.expanded !== '1' || Date.now() < suppressUntil) {
-                // If collapsed, expand instead of acting
-                if (parentLi.dataset.expanded !== '1') {
-                  parentLi.dataset.expanded = '1';
-                  parentLi.dataset.actionSuppressUntil = String(Date.now() + CLICK_EXPAND_SUPPRESS_MS);
-                  positionRadialStack(parentLi, grimoireState.players[i].reminders.length);
-                }
-                return;
-              }
-            }
-            // No-op on desktop; use hover edit icon instead
-          };
-          // Desktop hover actions on text reminders
-          if (!isTouchDevice) {
-            const editBtn2 = document.createElement('div');
-            editBtn2.className = 'reminder-action edit';
-            editBtn2.title = 'Edit';
-            editBtn2.innerHTML = '<i class="fa-solid fa-pen"></i>';
-            editBtn2.addEventListener('click', (e) => {
-              e.stopPropagation();
-              try { e.preventDefault(); } catch (_) { }
-              const parentLi = editBtn2.closest('li');
-              if (parentLi) {
-                const suppressUntil = parseInt(parentLi.dataset.actionSuppressUntil || '0', 10);
-                if (parentLi.dataset.expanded !== '1' || Date.now() < suppressUntil) {
-                  if (parentLi.dataset.expanded !== '1') {
-                    parentLi.dataset.expanded = '1';
-                    parentLi.dataset.actionSuppressUntil = String(Date.now() + CLICK_EXPAND_SUPPRESS_MS);
-                    positionRadialStack(parentLi, grimoireState.players[i].reminders.length);
-                  }
-                  return;
-                }
-              }
-              const current = grimoireState.players[i].reminders[idx]?.label || grimoireState.players[i].reminders[idx]?.value || '';
-              const next = prompt('Edit reminder', current);
-              if (next !== null) {
-                grimoireState.players[i].reminders[idx].value = next;
-                if (grimoireState.players[i].reminders[idx].label !== undefined) {
-                  grimoireState.players[i].reminders[idx].label = next;
-                }
-                updateGrimoire();
-                saveAppState();
-              }
-            });
-            reminderEl.appendChild(editBtn2);
-
-            const delBtn2 = document.createElement('div');
-            delBtn2.className = 'reminder-action delete';
-            delBtn2.title = 'Delete';
-            delBtn2.innerHTML = '<i class="fa-solid fa-trash"></i>';
-            delBtn2.addEventListener('click', (e) => {
-              e.stopPropagation();
-              try { e.preventDefault(); } catch (_) { }
-              const parentLi = delBtn2.closest('li');
-              if (parentLi) {
-                const suppressUntil = parseInt(parentLi.dataset.actionSuppressUntil || '0', 10);
-                if (parentLi.dataset.expanded !== '1' || Date.now() < suppressUntil) {
-                  if (parentLi.dataset.expanded !== '1') {
-                    parentLi.dataset.expanded = '1';
-                    parentLi.dataset.actionSuppressUntil = String(Date.now() + CLICK_EXPAND_SUPPRESS_MS);
-                    positionRadialStack(parentLi, grimoireState.players[i].reminders.length);
-                  }
-                  return;
-                }
-              }
-              grimoireState.players[i].reminders.splice(idx, 1);
-              updateGrimoire();
-              saveAppState();
-            });
-            reminderEl.appendChild(delBtn2);
-          }
-          // Touch long-press for reminder context menu
-          if (isTouchDevice) {
-            const onPressStart2 = (e) => {
-              try { e.preventDefault(); } catch (_) { }
-              clearTimeout(grimoireState.longPressTimer);
-              try { reminderEl.classList.add('press-feedback'); } catch (_) { }
-              const x = (e && (e.clientX || (e.touches && e.touches[0] && e.touches[0].clientX))) || 0;
-              const y = (e && (e.clientY || (e.touches && e.touches[0] && e.touches[0].clientY))) || 0;
-              grimoireState.longPressTimer = setTimeout(() => {
-                try { reminderEl.classList.remove('press-feedback'); } catch (_) { }
-                showReminderContextMenu(x, y, i, idx);
-              }, 600);
-            };
-            const onPressEnd2 = () => { clearTimeout(grimoireState.longPressTimer); try { reminderEl.classList.remove('press-feedback'); } catch (_) { } };
-            reminderEl.addEventListener('pointerdown', onPressStart2);
-            reminderEl.addEventListener('pointerup', onPressEnd2);
-            reminderEl.addEventListener('pointercancel', onPressEnd2);
-            reminderEl.addEventListener('pointerleave', onPressEnd2);
-            reminderEl.addEventListener('touchstart', onPressStart2, { passive: false });
-            reminderEl.addEventListener('touchend', onPressEnd2);
-            reminderEl.addEventListener('touchcancel', onPressEnd2);
-            reminderEl.addEventListener('contextmenu', (e) => { try { e.preventDefault(); } catch (_) { } });
-          }
-
-          remindersDiv.appendChild(reminderEl);
-        }
-      });
-
-      // After rendering, position all reminders and the plus button in a radial stack
-      positionRadialStack(li, player.reminders.length);
-    });
-
-    // Position info icons after updating grimoire
-    if ('ontouchstart' in window) {
-      positionInfoIcons();
-    }
-  }
 
   // Arrange reminders and plus button along the line from token center to circle center
   function positionRadialStack(li, count) { positionRadialStackLayout(li, count, grimoireState.players); }
@@ -1454,9 +803,9 @@ document.addEventListener('DOMContentLoaded', () => {
     if (grimoireState.selectedPlayerIndex > -1) {
       grimoireState.players[grimoireState.selectedPlayerIndex].character = roleId;
       console.log(`Assigned character ${roleId} to player ${grimoireState.selectedPlayerIndex}`);
-      updateGrimoire();
+      updateGrimoire({ grimoireState });
       characterModal.style.display = 'none';
-      saveAppState();
+      saveAppState({ grimoireState });
     }
   }
 
@@ -1483,8 +832,8 @@ document.addEventListener('DOMContentLoaded', () => {
     } else if (reminderIndex > -1) {
       grimoireState.players[playerIndex].reminders.splice(reminderIndex, 1);
     }
-    updateGrimoire();
-    saveAppState();
+    updateGrimoire({ grimoireState });
+    saveAppState({ grimoireState });
     textReminderModal.style.display = 'none';
   };
 
@@ -1503,10 +852,6 @@ document.addEventListener('DOMContentLoaded', () => {
     const content = reminderTokenModal.querySelector('.modal-content');
     if (content && !content.contains(e.target)) { reminderTokenModal.style.display = 'none'; }
   });
-
-  // createCurvedLabelSvg moved to ui/svg.js
-
-  // createDeathRibbonSvg moved to ui/svg.js
 
   function openReminderTokenModal(playerIndex) {
     grimoireState.selectedPlayerIndex = playerIndex;
@@ -1593,8 +938,8 @@ document.addEventListener('DOMContentLoaded', () => {
             label = input;
           }
           grimoireState.players[grimoireState.selectedPlayerIndex].reminders.push({ type: 'icon', id: token.id, image: token.image, label, rotation: 0 });
-          updateGrimoire();
-          saveAppState();
+          updateGrimoire({ grimoireState });
+          saveAppState({ grimoireState });
           reminderTokenModal.style.display = 'none';
         };
         tokenEl.addEventListener('click', handleSelect);
@@ -1622,7 +967,7 @@ document.addEventListener('DOMContentLoaded', () => {
     resizeObserver = new ResizeObserver((entries) => {
       if (grimoireState.players.length > 0) {
         console.log('Container resized, repositioning players...');
-        requestAnimationFrame(repositionPlayers);
+        requestAnimationFrame(() => repositionPlayers(grimoireState.players));
       }
     });
 
@@ -1639,7 +984,7 @@ document.addEventListener('DOMContentLoaded', () => {
       resizeTimeout = setTimeout(() => {
         if (grimoireState.players.length > 0) {
           console.log('Window resized, repositioning players...');
-          requestAnimationFrame(repositionPlayers);
+          requestAnimationFrame(() => repositionPlayers(grimoireState.players));
         }
       }, 250);
     });
@@ -1648,7 +993,7 @@ document.addEventListener('DOMContentLoaded', () => {
   // Also reposition players when the page becomes visible (in case of tab switching)
   document.addEventListener('visibilitychange', () => {
     if (!document.hidden && grimoireState.players.length > 0) {
-      requestAnimationFrame(repositionPlayers);
+      requestAnimationFrame(() => repositionPlayers(grimoireState.players));
     }
   });
 
@@ -1774,11 +1119,6 @@ document.addEventListener('DOMContentLoaded', () => {
     }
   }
 
-  // Helper to get role by id respecting traveller toggle
-  function getRoleById(roleId) {
-    return grimoireState.allRoles[roleId] || grimoireState.baseRoles[roleId] || grimoireState.extraTravellerRoles[roleId] || null;
-  }
-
   // Sidebar resizer
   initSidebarResize(sidebarResizer, sidebarEl);
 
@@ -1789,7 +1129,6 @@ document.addEventListener('DOMContentLoaded', () => {
     sidebarBackdrop,
     sidebarEl,
     sidebarResizer,
-    prefersOverlaySidebar,
     isTouchDevice,
     repositionPlayers
   });
@@ -1800,7 +1139,7 @@ document.addEventListener('DOMContentLoaded', () => {
   renderGrimoireHistory({ grimoireHistoryList });
 
   // Restore previous session (script and grimoire)
-  loadAppState();
+  loadAppState({ grimoireState, grimoireHistoryList, processScriptData, repositionPlayers, openCharacterModal, showPlayerContextMenu, openReminderTokenModal, openTextReminderModal });
 
   // In-app tour
   initInAppTour();
